add schema, mongooos, connection, joi
