+++ schemas/contacts-schemas.js
@@ -1,0 +1,18 @@
+import Joi from "joi";
+
+const contactsAddSchema = Joi.object({
+  name: Joi.string().required().messages({
+    "any.required": `"name" must be exist`,
+    "string.empty": `"name" cannot be empty`,
+  }),
+  email: Joi.string().required().messages({
+    "any.required": `"email" must be exist`,
+    "string.empty": `"email" cannot be empty`,
+  }),
+  phone: Joi.string().required().messages({
+    "any.required": `"phone" must be exist`,
+    "string.empty": `"phone" cannot be empty`,
+  }),
+});
+
+export default { contactsAddSchema };
