+++ server.js
@@ -1,5 +1,18 @@
 import app from "./app.js";
+import mongoose from "mongoose";
+import dotenv from "dotenv";
 
-app.listen(3000, () => {
-  console.log("Server running. Use our API on port: 3000");
-});
+dotenv.config();
+const { DB_HOST, PORT } = process.env;
+
+mongoose
+  .connect(DB_HOST)
+  .then(
+    app.listen(PORT, () => {
+      console.log(`Database connection successful on port: ${PORT}`);
+    })
+  )
+  .catch((error) => {
+    console.log(error.message);
+    process.exit(1);
+  });
